% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{H2c}
\alias{H2c}
\title{Kernel regression method}
\usage{
H2c(x, y, kernel='gaussian', d=1, sigma='default', bw='default', wt=F)
}
\arguments{
\item{x}{a numeric vector/matrix}

\item{y}{a numeric vector/matrix}

\item{kernel}{a kernel to use for x, 'gaussian' or 'distance'}

\item{d}{exponent on distance in (0,2], if a 'distance' kernel is used for x}

\item{sigma}{bandwidth, if a 'gaussian' kernel is used for x, 
default is heuristic median pairwise distances of x}

\item{bw}{bandwidth of the gaussian smoothing kernel applied on y,
bandwidths suggested by Silverman (1986) are used unless otherwise specified.}

\item{wt}{use weight function or not, if TRUE, f^2 is used by default}
}
\description{
compute the ECCFIC between continuous x and y, 
             need to load the library kernlab
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100)
result <- H2c(x,y)
}
