% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{H2d}
\alias{H2d}
\title{Slicing method}
\usage{
H2d(x, y, kernel='gaussian', d=1, sigma='default')
}
\arguments{
\item{x}{a numeric vector/matrix}

\item{y}{a numeric vector}

\item{kernel}{a kernel to use for x, 'gaussian' or 'distance'}

\item{d}{exponent on distance in (0,2], if a 'distance' kernel is used}

\item{sigma}{bandwidth, if a 'gaussian' kernel is used, 
default is heuristic median pairwise distances of x}
}
\description{
compute the ECCFIC between x and y, where y is categorical
}
\examples{
y <- rep(1:4, 25)
x <- y + rnorm(100)
result <- H2d(x,y)
}
