% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{ptest_H2c}
\alias{ptest_H2c}
\title{Permutation test with kernel regression estimator}
\usage{
ptest_H2c(
  x,
  y,
  kernel = "gaussian",
  d = 1,
  sigma = "default",
  bw = "default",
  wt = F,
  B
)
}
\arguments{
\item{B}{number of replicates}
}
\description{
Permutation test with kernel regression estimator
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100)
result <- ptest_H2c(x,y,B=199)
}
