% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{ptest_H2d}
\alias{ptest_H2d}
\title{Permutation test with slicing estimator}
\usage{
ptest_H2d(x, y, kernel = "gaussian", d = 1, sigma = "default", B)
}
\arguments{
\item{B}{number of replicates}
}
\description{
Permutation test with slicing estimator
}
\examples{
y <- rep(1:4, 25)
x <- y + rnorm(100)
result <- ptest_H2d(x,y,B=199)
}
